+++ webapp/locales/en_US.json
@@ -319,6 +319,7 @@
   "Email successfully sent": "Email successfully sent",
   "Emails": "Emails",
   "Encoding": "Encoding",
+  "Enter the file name!": "Enter the file name!",
   "Enter the table name!": "Enter the table name!",
   "Error!": "Error!",
   "Error": "Error",
